improve memory debugging functions

+++ aoo/src/aoo.cpp
@@ -76,14 +76,14 @@ int32_t get_random_id(){
 namespace aoo {
 
 #if AOO_DEBUG_MEMORY
-std::atomic<int64_t> total_memory{0};
+std::atomic<ptrdiff_t> total_memory{0};
 #endif
 
 static AooAllocator g_allocator {
     [](size_t n, void *){
     #if AOO_DEBUG_MEMORY
-        auto total = total_memory.fetch_add(n, std::memory_order_relaxed) + n;
-        fprintf(stderr, "allocate %d bytes (total: %d)\n", n, total);
+        auto total = total_memory.fetch_add(n, std::memory_order_relaxed) + (ptrdiff_t)n;
+        fprintf(stderr, "allocate %zu bytes (total: %zd)\n", n, total);
         fflush(stderr);
     #endif
         return operator new(n);
@@ -91,8 +91,8 @@ static AooAllocator g_allocator {
     nullptr,
     [](void *ptr, size_t n, void *){
     #if AOO_DEBUG_MEMORY
-        auto total = total_memory.fetch_sub(n, std::memory_order_relaxed) - n;
-        fprintf(stderr, "deallocate %d bytes (total: %d)\n", n, total);
+        auto total = total_memory.fetch_sub(n, std::memory_order_relaxed) - (ptrdiff_t)n;
+        fprintf(stderr, "deallocate %zu bytes (total: %zd)\n", n, total);
         fflush(stderr);
     #endif
         operator delete(ptr);
